refactor(backend): extract prompt crud module
