+++ tests/backend/security/test_organization_filtering.py
@@ -12,6 +12,7 @@ from sqlalchemy.orm import Session
 
 from rhesis.backend.app import crud, models
 from rhesis.backend.app.crud import model as model_crud
+from rhesis.backend.app.crud import prompt as prompt_crud
 from rhesis.backend.app.crud import task as task_crud
 from rhesis.backend.app.crud import test_result as test_result_crud
 from rhesis.backend.app.crud.metric import create_metric, get_metric
@@ -97,7 +98,7 @@ class TestCrudOrganizationFiltering:
         from tests.backend.routes.fixtures.data_factories import PromptDataFactory
 
         prompt_data = PromptDataFactory.minimal_data()
-        prompt = crud.create_prompt(
+        prompt = prompt_crud.create_prompt(
             db=test_db, prompt=prompt_data, organization_id=str(org1.id), user_id=str(user1.id)
         )
 
@@ -145,7 +146,7 @@ class TestCrudOrganizationFiltering:
         )
 
         prompt_data = PromptDataFactory.minimal_data()
-        prompt = crud.create_prompt(
+        prompt = prompt_crud.create_prompt(
             db=test_db, prompt=prompt_data, organization_id=str(org1.id), user_id=str(user1.id)
         )
 
@@ -235,7 +236,7 @@ class TestCrudOrganizationFiltering:
         )
 
         prompt_data = PromptDataFactory.minimal_data()
-        prompt = crud.create_prompt(
+        prompt = prompt_crud.create_prompt(
             db=test_db, prompt=prompt_data, organization_id=str(org1.id), user_id=str(user1.id)
         )
 
@@ -377,12 +378,12 @@ class TestCrudOrganizationFiltering:
         from tests.backend.routes.fixtures.data_factories import PromptDataFactory
 
         prompt_data = PromptDataFactory.minimal_data()
-        prompt = crud.create_prompt(
+        prompt = prompt_crud.create_prompt(
             db=test_db, prompt=prompt_data, organization_id=str(org1.id), user_id=str(user1.id)
         )
 
         # User from org1 should be able to access the prompt
-        result_org1 = crud.get_prompt(
+        result_org1 = prompt_crud.get_prompt(
             test_db, prompt.id, organization_id=str(org1.id), user_id=str(user1.id)
         )
         assert result_org1 is not None
@@ -390,7 +391,7 @@ class TestCrudOrganizationFiltering:
         assert str(result_org1.organization_id) == str(org1.id)
 
         # User from org2 should NOT be able to access the prompt
-        result_org2 = crud.get_prompt(
+        result_org2 = prompt_crud.get_prompt(
             test_db, prompt.id, organization_id=str(org2.id), user_id=str(user2.id)
         )
         assert result_org2 is None
@@ -613,7 +614,7 @@ class TestCrudParameterValidation:
             ("get_test_result", test_result_crud.get_test_result),
             ("get_test_run", get_test_run),
             ("get_endpoint", crud.get_endpoint),
-            ("get_prompt", crud.get_prompt),
+            ("get_prompt", prompt_crud.get_prompt),
             ("get_model", model_crud.get_model),
             ("get_metric", get_metric),
         ]
